docs: update loading patterns (#868)
## Description

More examples and guidelines on how to use our loading indicators
(Spinner and Skeleton)

## Changes

create `LoadingExamples.tsx`
create examples for Skeleton, full screen loading, inline loading,
button loading
add css to custom.css
update guidelines on `loading-pattern.mdx`


## Additional Information

The overlay for full screen spinner should either have a token or a
variant in the component for consistency
Inline spinner needs a container of some sort, I had to make my own now
to align it properly with the field. Maybe an inline-spinner variant?
The small spinner is 16px. Our icons are 20, so it should probably be 20
instead. This is fixed in #880

---------

Co-authored-by: derpbravely <197733689+derpbravely@users.noreply.github.com>